+++ app/src/main/java/com/example/data/local/TransactionEntity.kt
@@ -2,6 +2,7 @@ package com.example.data.local
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
+import java.math.BigDecimal
 
 enum class TransactionType {
     DEPOSIT, TRANSFER, SWAP, EXPENSE
@@ -12,7 +13,7 @@ data class TransactionEntity(
     @PrimaryKey(autoGenerate = true)
     val id: Long = 0,
     val title: String,
-    val amount: Double, // Amount in Toman
+    val amount: BigDecimal, // Amount in Toman
     val type: TransactionType,
     val category: String,
     val accountId: Long? = null,
