+++ app/src/main/java/com/example/domain/model/BankAccount.kt
@@ -1,9 +1,11 @@
 package com.example.domain.model
 
+import java.math.BigDecimal
+
 data class BankAccount(
     val id: Long,
     val name: String,
     val bankName: String,
-    val currentBalance: Double,
+    val currentBalance: BigDecimal,
     val colorHex: String
 )
