+++ app/src/main/java/com/example/domain/model/MarketAsset.kt
@@ -1,5 +1,7 @@
 package com.example.domain.model
 
+import java.math.BigDecimal
+
 enum class MarketAssetType {
     GOLD, COIN, CURRENCY, STOCK, COMMODITY, CRYPTO, CASH
 }
@@ -13,10 +15,10 @@ data class MarketAsset(
     val symbol: String,
     val name: String,
     val type: MarketAssetType,
-    val price: Double,          // Base unit depends on type (usually Rial/Toman)
-    val previousPrice: Double,
-    val change: Double,
-    val changePercent: Double,
+    val price: BigDecimal,          // Base unit depends on type (usually Rial/Toman)
+    val previousPrice: BigDecimal,
+    val change: BigDecimal,
+    val changePercent: BigDecimal,
     val timestamp: Long,
     val source: String
 )
